+++ EcoMarket/Scenes/TabBar/Cart/CartTheme1/CartModel.swift
@@ -16,6 +16,9 @@ struct CartModel {
     static var products: [CartModel] = [
         .init(name: "Vado Odelle Dress", brand: "Roller Rabbit", price: "$198.00", image: "cart/image1"),
         .init(name: "Clean 90 Triole Snakers", brand: "Axel Arigato", price: "$245.00", image: "cart/image2"),
+        .init(name: "Daypack Backpack", brand: "Herschel Supply Co.", price: "$40.00", image: "cart/image3"),
+        .init(name: "Daypack Backpack", brand: "Herschel Supply Co.", price: "$40.00", image: "cart/image3"),
+        .init(name: "Daypack Backpack", brand: "Herschel Supply Co.", price: "$40.00", image: "cart/image3"),
         .init(name: "Daypack Backpack", brand: "Herschel Supply Co.", price: "$40.00", image: "cart/image3")
     ]
 }
